+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
-Nothing yet.
+### Added
+
+- Support for aliased subcommands, starting with delete = rm and push = deploy.
 
 
 ## [0.4.1] - 2018-03-12